discount and availability for each product added to returned cart, products now are returned ordered by time they were added to cart

+++ src/main/java/com/heroku/spacey/dao/ProductDao.java
@@ -29,7 +29,7 @@ public interface ProductDao {
 
     int updateSizeToProduct(SizeToProduct sizeToProduct);
 
-    double getAmount(Long sizeId, Long productId);
+    int getAmount(Long sizeId, Long productId);
 
     ProductForUnauthorizedCart getProductByIdAndSize(Long productId, Long sizeId);
 }
